requirements/base.txt
Django==3.2.25
djangorestframework==3.15.1
drf-yasg==1.21.10
openai==2.17.0
django-environ
